+++ lib/faastruby/cli/commands/function/update_context.rb
@@ -16,8 +16,8 @@ module FaaStRuby
         def run
           spinner = spin("Uploading context data to '#{@workspace_name}'...")
           workspace = FaaStRuby::Workspace.new(name: @workspace_name)
-          function = FaaStRuby::Function.new(name: @function_name, workspace: workspace, context: @options['data'])
-          function.update
+          function = FaaStRuby::Function.new(name: @function_name, workspace: workspace)
+          function.update(new_context: @options['data'])
           if function.errors.any?
             spinner.stop('Failed :(')
             FaaStRuby::CLI.error(function.errors)