+++ spec/faastruby/api_spec.rb
@@ -1,0 +1,119 @@
+require 'tempfile'
+RSpec.describe FaaStRuby::API do
+  before do
+    FaaStRuby.configure do |config|
+      config.api_key = 'API_KEY'
+      config.api_secret = 'API_SECRET'
+    end
+    @api = FaaStRuby::API.new
+    @fake_api = FaaStRuby::Tests::API.new
+  end
+  describe '#initialize' do
+    it 'sets @api_url, @credentials and @headers' do
+      expect(@api.api_url).to eq("#{FAASTRUBY_HOST}/v2")
+      expect(@api.credentials).to eq({'API-KEY' => 'API_KEY', 'API-SECRET' => 'API_SECRET'})
+      expect(@api.headers).to eq({content_type: :json, accept: :json, 'API-KEY' => 'API_KEY', 'API-SECRET' => 'API_SECRET'})
+    end
+  end
+
+  describe '#create_workspace' do
+    before do
+      @url_path = "#{FAASTRUBY_HOST}/v2/workspaces"
+      @fake_api.stub_create_workspace(name: 'test-workspace')
+      @api_response = @api.create_workspace(workspace_name: 'test-workspace')
+    end
+    it 'makes a request' do
+      expect(WebMock).to have_requested(:post, @url_path)
+    end
+    it 'returns a struct with body, errors and status code' do
+      expect(@api_response).to have_attributes(body: @fake_api.create_workspace_response, code: 200, errors: [])
+    end
+  end
+
+  describe '#destroy_workspace' do
+    before do
+      @url_path = "#{FAASTRUBY_HOST}/v2/workspaces/test-workspace"
+      @fake_api.stub_destroy_workspace(name: 'test-workspace')
+      @api_response = @api.destroy_workspace('test-workspace')
+    end
+    it 'makes a request' do
+      expect(WebMock).to have_requested(:delete, @url_path)
+    end
+    it 'returns a struct with body, errors and status code' do
+      expect(@api_response).to have_attributes(body: {}, code: 200, errors: [])
+    end
+  end
+
+  describe '#get_workspace_info' do
+    before do
+      @url_path = "#{FAASTRUBY_HOST}/v2/workspaces/test-workspace"
+      @fake_api.stub_get_workspace(name: 'test-workspace')
+      @api_response = @api.get_workspace_info('test-workspace')
+    end
+    it 'makes a request' do
+      expect(WebMock).to have_requested(:get, @url_path)
+    end
+    it 'returns a struct with body, errors and status code' do
+      expect(@api_response).to have_attributes(body: @fake_api.get_workspace_response, code: 200, errors: [])
+    end
+  end
+
+  describe '#deploy' do
+    before do
+      @url_path = "#{FAASTRUBY_HOST}/v2/workspaces/test-workspace/deploy"
+      @fake_api.stub_deploy_workspace(name: 'test-workspace')
+      @file = Tempfile.new('file.zip')
+      @api_response = @api.deploy(workspace_name: 'test-workspace', package: @file)
+    end
+    it 'makes a request' do
+      expect(WebMock).to have_requested(:post, @url_path)
+    end
+    it 'returns a struct with body, errors and status code' do
+      expect(@api_response).to have_attributes(body: {}, code: 200, errors: [])
+    end
+  end
+
+  describe '#delete_from_workspace' do
+    before do
+      @url_path = "#{FAASTRUBY_HOST}/v2/workspaces/test-workspace/functions/test-function"
+      @fake_api.stub_delete_from_workspace(workspace_name: 'test-workspace', function_name: 'test-function')
+      @api_response = @api.delete_from_workspace(workspace_name: 'test-workspace', function_name: 'test-function')
+    end
+    it 'makes a request' do
+      expect(WebMock).to have_requested(:delete, @url_path)
+    end
+    it 'returns a struct with body, errors and status code' do
+      expect(@api_response).to have_attributes(body: {}, code: 200, errors: [])
+    end
+  end
+
+  describe '#run' do
+    before do
+      @url_path = "#{FAASTRUBY_HOST}/test-workspace/test-function"
+      @fake_api.stub_run_function(function_name: 'test-function', workspace_name: 'test-workspace', method: :post, query: '?foo=bar')
+      @api_response = @api.run(workspace_name: 'test-workspace', function_name: 'test-function', payload: 'Example function', method: 'post', headers: {'Example' => 'Header'}, query: '?foo=bar')
+    end
+    it 'makes a request' do
+      expect(WebMock).to have_requested(:post, @url_path).with(query: {'foo' => 'bar'})
+    end
+    it 'returns a rest-client response' do
+      expect(@api_response.body).to eq('Function response')
+      expect(@api_response.code).to eq(200)
+    end
+
+    describe '#update_function_context' do
+      before do
+        @url_path = "#{FAASTRUBY_HOST}/v2/workspaces/test-workspace/functions/test-function"
+        @fake_api.stub_update_function(workspace_name: 'test-workspace', function_name: 'test-function', context: {'context' => 'new context'})
+        @api_response = @api.update_function_context(workspace_name: 'test-workspace', function_name: 'test-function', payload: {'context' => 'new context'})
+      end
+      it 'makes a request' do
+        expect(WebMock).to have_requested(:patch, @url_path)
+      end
+      it 'returns a struct with body, errors and status code' do
+        expect(@api_response).to have_attributes(body: @fake_api.function_response('new context'), code: 200, errors: [])
+      end
+    end
+
+  end
+end